process handling: portability: introduce directory-like abstraction
